+++ language-server/modules/langserver-core/src/main/java/org/ballerinalang/langserver/common/utils/CommonUtil.java
@@ -253,267 +253,6 @@ public class CommonUtil {
     }
 
     /**
-     * Get the default value for the given BType.
-     *
-     * @param bType  Type descriptor to get the default value
-     * @param offset snippet offset.
-     * @return {@link String}   Default value as a String
-     */
-    public static Optional<String> getDefaultValueForType(TypeSymbol bType, int offset) {
-        return getDefaultValueForType(bType, true, offset);
-    }
-
-    public static Optional<String> getDefaultPlaceholderForType(TypeSymbol bType) {
-        return getDefaultValueForType(bType)
-                .map(defaultValue -> defaultValue.replace("}", "\\}"));
-    }
-
-    /**
-     * Get the default value for the given BType.
-     *
-     * @param bType Type descriptor to get the default value
-     * @return {@link String}   Default value as a String
-     */
-    public static Optional<String> getDefaultValueForType(TypeSymbol bType) {
-        return getDefaultValueForType(bType, false, -1);
-    }
-
-    private static Optional<String> getDefaultValueForType(TypeSymbol bType, boolean isSnippet, int offset) {
-        String valueString;
-        if (bType == null) {
-            return Optional.empty();
-        }
-
-        TypeSymbol rawType = getRawType(bType);
-        TypeDescKind typeKind = rawType.typeKind();
-        switch (typeKind) {
-            case TUPLE:
-                TupleTypeSymbol tupleType = (TupleTypeSymbol) rawType;
-                List<String> memberDefaultValues = tupleType.memberTypeDescriptors().stream()
-                        .map(member -> getDefaultValueForTypeDescKind(member).orElse(""))
-                        .collect(Collectors.toList());
-
-                if (memberDefaultValues.isEmpty()) {
-                    valueString = isSnippet ? "[${" + offset + "}]" : "[]";
-                    break;
-                }
-                List<String> memberSnippets = new ArrayList<>();
-                int memberOffset = offset;
-                for (String value : memberDefaultValues) {
-                    memberSnippets.add(isSnippet ? generateSnippetEntry(value, memberOffset) : value);
-                    memberOffset++;
-                }
-                valueString = "[" + String.join(", ", memberSnippets) + "]";
-                break;
-            case ARRAY:
-                // Filler value of an array is []
-                ArrayTypeSymbol arrayType = (ArrayTypeSymbol) rawType;
-                if (arrayType.memberTypeDescriptor().typeKind() == TypeDescKind.ARRAY) {
-                    String value = getDefaultValueForTypeDescKind(arrayType.memberTypeDescriptor())
-                            .orElse("");
-                    valueString = "[" + (isSnippet ? generateSnippetEntry(value, offset) : value) + "]";
-                    break;
-                }
-                valueString = isSnippet ? "[${" + offset + "}]" : "[]";
-                break;
-            case RECORD:
-                // TODO: Here we have disregarded the formatting of the record fields. Need to consider that in future
-                RecordTypeSymbol recordTypeSymbol = (RecordTypeSymbol) rawType;
-                List<String> fieldInsertText = new ArrayList<>();
-                int count = offset;
-                valueString = "{";
-                List<RecordFieldSymbol> mandatoryFieldSymbols = getMandatoryRecordFields(recordTypeSymbol).stream()
-                        .filter(recordFieldSymbol -> recordFieldSymbol.getName().isPresent())
-                        .collect(Collectors.toList());
-                for (RecordFieldSymbol mandatoryField : mandatoryFieldSymbols) {
-                    String value = getDefaultValueForTypeDescKind(CommonUtil.getRawType(mandatoryField
-                            .typeDescriptor())).orElse("");
-                    fieldInsertText.add(mandatoryField.getName().get() + ": "
-                            + (isSnippet ? generateSnippetEntry(value, count) : value));
-                    count++;
-                }
-                valueString += String.join(", ", fieldInsertText);
-                valueString += "}";
-                break;
-            case OBJECT:
-                ObjectTypeSymbol objectTypeSymbol = (ObjectTypeSymbol) rawType;
-                if (objectTypeSymbol.kind() == SymbolKind.CLASS) {
-                    ClassSymbol classSymbol = (ClassSymbol) objectTypeSymbol;
-                    if (classSymbol.initMethod().isPresent()) {
-                        List<ParameterSymbol> params = classSymbol.initMethod().get().typeDescriptor().params().get();
-                        int snippetIndex = offset;
-                        List<String> paramSnippets = new ArrayList<>();
-                        for (ParameterSymbol param : params) {
-                            String value =
-                                    getDefaultValueForTypeDescKind(CommonUtil.getRawType(param.typeDescriptor()))
-                                            .orElse("");
-                            paramSnippets.add(isSnippet ? generateSnippetEntry(value, snippetIndex) : value);
-                            snippetIndex++;
-                        }
-                        valueString = "new (" + String.join(", ", paramSnippets) + ")";
-                    } else {
-                        valueString = isSnippet ? "${" + offset + ":new ()}" : "new ()";
-                    }
-                } else {
-                    valueString = isSnippet ? "${" + offset + ":object {}}" : "object {}";
-                }
-                break;
-            case UNION:
-                List<TypeSymbol> members =
-                        new ArrayList<>(((UnionTypeSymbol) rawType).memberTypeDescriptors());
-                List<TypeSymbol> nilMembers = members.stream()
-                        .filter(member -> member.typeKind() == TypeDescKind.NIL).collect(Collectors.toList());
-                if (nilMembers.isEmpty()) {
-                    valueString = getDefaultValueForTypeDescKind(CommonUtil.getRawType(members.get(0))).orElse("");
-                    valueString = isSnippet ? generateSnippetEntry(valueString, offset) : valueString;
-                } else {
-                    return Optional.of(isSnippet ? generateSnippetEntry("()", offset) : "()");
-                }
-                break;
-            case INTERSECTION:
-                TypeSymbol effectiveType = ((IntersectionTypeSymbol) rawType).effectiveTypeDescriptor();
-                effectiveType = getRawType(effectiveType);
-                if (effectiveType.typeKind() == TypeDescKind.INTERSECTION) {
-                    // Right now, intersection types can only have readonly and another type only. Therefore, not doing 
-                    // further checks here. Get the member type from intersection which is not readonly and get its 
-                    // default value
-                    valueString = isSnippet ? generateSnippetEntry("()", offset) : "()";
-                    Optional<TypeSymbol> memberType = ((IntersectionTypeSymbol) effectiveType)
-                            .memberTypeDescriptors().stream()
-                            .filter(typeSymbol -> typeSymbol.typeKind() != TypeDescKind.READONLY)
-                            .findAny();
-                    if (memberType.isPresent()) {
-                        valueString =
-                                getDefaultValueForTypeDescKind(CommonUtil.getRawType(memberType.get())).orElse("");
-                        valueString = isSnippet ? generateSnippetEntry(valueString, offset) : valueString;
-                    }
-                } else {
-                    valueString = getDefaultValueForTypeDescKind(CommonUtil.getRawType(effectiveType))
-                            .orElse("");
-                    valueString = isSnippet ? generateSnippetEntry(valueString, offset) : valueString;
-                }
-                break;
-            case TABLE:
-                TypeSymbol rowType = ((TableTypeSymbol) rawType).rowTypeParameter();
-                String rowValue = getDefaultValueForTypeDescKind(rowType).orElse("");
-                valueString = "table [" + (isSnippet ? generateSnippetEntry(rowValue, offset) : rowValue) + "]";
-                break;
-            case ERROR:
-                TypeSymbol errorType = CommonUtil.getRawType(((ErrorTypeSymbol) rawType).detailTypeDescriptor());
-                StringBuilder errorString = new StringBuilder("error (\"\"");
-                if (errorType.typeKind() == TypeDescKind.RECORD) {
-                    List<RecordFieldSymbol> mandatoryFields = getMandatoryRecordFields((RecordTypeSymbol) errorType);
-                    if (!mandatoryFields.isEmpty()) {
-                        errorString.append(", ");
-                        List<String> detailFieldSnippets = new ArrayList<>();
-                        int index = offset;
-                        for (RecordFieldSymbol field : mandatoryFields) {
-                            String defValue =
-                                    getDefaultValueForTypeDescKind(CommonUtil.getRawType(field.typeDescriptor()))
-                                            .orElse("");
-                            detailFieldSnippets.add(field.getName().get()
-                                    + " = " + (isSnippet ? generateSnippetEntry(defValue, index) : defValue));
-                            index++;
-                        }
-                        errorString.append(String.join(", ", detailFieldSnippets));
-                    }
-                }
-                errorString.append(")");
-                valueString = errorString.toString();
-                break;
-            case SINGLETON:
-                valueString = rawType.signature();
-                valueString = isSnippet ? generateSnippetEntry(valueString, offset) : valueString;
-                break;
-            default:
-                Optional<String> value = getDefaultValueForTypeDescKind(typeKind);
-                if (value.isEmpty()) {
-                    return value;
-                }
-                return isSnippet ? Optional.of(generateSnippetEntry(value.get(), offset)) : value;
-        }
-
-        return Optional.of(valueString);
-    }
-
-    private static String generateSnippetEntry(String value, int offset) {
-        return "${" + offset + ":" + value + "}";
-    }
-
-    /**
-     * Used to get the default values for a {@link TypeDescKind}. {@link #getDefaultValueForType(TypeSymbol)}
-     * is preferred over this function. This function should be used as a compliment to .
-     *
-     * @param typeKind Type desc kind
-     * @return Optional default value
-     * @see #getDefaultValueForType(TypeSymbol)
-     */
-    public static Optional<String> getDefaultValueForTypeDescKind(TypeDescKind typeKind) {
-        String defaultValue = null;
-        switch (typeKind) {
-            case FLOAT:
-                defaultValue = Float.toString(0);
-                break;
-            case BOOLEAN:
-                defaultValue = Boolean.toString(false);
-                break;
-            case RECORD:
-            case MAP:
-                defaultValue = "{}";
-                break;
-            case STREAM:
-                defaultValue = "new ()";
-                break;
-            case XML:
-                defaultValue = "xml ``";
-                break;
-            case DECIMAL:
-                defaultValue = Integer.toString(0);
-                break;
-            case ARRAY:
-                defaultValue = "[]";
-                break;
-            case SINGLETON:
-                defaultValue = "\"\"";
-                break;
-            case ERROR:
-                defaultValue = "error(\"\")";
-                break;
-            default:
-                if (typeKind.isIntegerType()) {
-                    defaultValue = Integer.toString(0);
-                    break;
-                } else if (typeKind.isStringType()) {
-                    defaultValue = "\"\"";
-                    break;
-                }
-        }
-        return Optional.ofNullable(defaultValue);
-    }
-
-    /**
-     * Used to get the default values for {@link TypeDescKind} of a {@link TypeSymbol}.
-     * {@link #getDefaultValueForType(TypeSymbol)} is preferred over this function.
-     * This function should be used as a compliment to .
-     *
-     * @param typeSymbol Type symbol
-     * @return Optional default value
-     * @see #getDefaultValueForType(TypeSymbol)
-     */
-    public static Optional<String> getDefaultValueForTypeDescKind(TypeSymbol typeSymbol) {
-        String defaultValue;
-        TypeDescKind typeKind = typeSymbol.typeKind();
-        switch (typeKind) {
-            case SINGLETON:
-                defaultValue = typeSymbol.signature();
-                break;
-            default:
-                defaultValue = getDefaultValueForTypeDescKind(typeKind).orElse(null);
-        }
-        return Optional.ofNullable(defaultValue);
-    }
-
-    /**
      * Get completion items list for struct fields.
      *
      * @param context Language server operation context
@@ -585,7 +324,7 @@ public class CommonUtil {
             for (Map.Entry<String, RecordFieldSymbol> entry : requiredFields.entrySet()) {
                 String fieldEntry = entry.getKey()
                         + PKG_DELIMITER_KEYWORD + " "
-                        + getDefaultValueForType(entry.getValue().typeDescriptor(), count).orElse(" ");
+                        + DefaultValueGenerationUtil.getDefaultValueForType(entry.getValue().typeDescriptor(), count).orElse(" ");
                 fieldEntries.add(fieldEntry);
                 count++;
             }
@@ -709,7 +448,7 @@ public class CommonUtil {
     public static String getRecordFieldCompletionInsertText(RecordFieldSymbol bField, int tabOffset) {
 
         StringBuilder insertText = new StringBuilder(bField.getName().get() + ": ");
-        insertText.append(getDefaultValueForType(bField.typeDescriptor(), tabOffset).orElse(" "));
+        insertText.append(DefaultValueGenerationUtil.getDefaultValueForType(bField.typeDescriptor(), tabOffset).orElse(" "));
         return insertText.toString();
     }
